Push down ElasticsearchService._summary and ._search
